+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/config/documentation/changelog/GenerateChangelogHtmlTask.kt
@@ -6,10 +6,12 @@ import org.gradle.api.file.ProjectLayout
 import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.model.ObjectFactory
 import org.gradle.api.provider.Property
+import org.gradle.api.tasks.CacheableTask
 import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.OutputFile
 import org.gradle.api.tasks.TaskAction
 
+@CacheableTask
 public abstract class GenerateChangelogHtmlTask
 @Inject
 constructor(objects: ObjectFactory, layout: ProjectLayout) : DefaultTask() {