+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Changed
 
+- `ValidatePlugins::enableStricterValidation` to `true`
+
 ### Deprecated
 
 ### Fixed